+++ src/solve/Solver.h
@@ -35,7 +35,7 @@
 #include "Integrator.h"
 #include "State.h"
 #include "debug.h"
-#include "csvwriter.h"
+//#include "csvwriter.h"
 #include "SimulationParameters.h"
 #include "Model.h"
